refactor(storage): hoist Postgres type mapping into shared mapPostgresType
PostgresTabularStorage and SupabaseTabularStorage each carried a ~110-line
mapTypeToSQL that was identical except that Postgres also emits pgvector
`vector(N)` columns. Extract one mapPostgresType(typeDef, { getNonNullType,
vectorTypeFor }) in @workglow/storage's sql/ module; the pgvector behavior
is injected via the optional vectorTypeFor hook (Postgres provides it,
Supabase omits it). Behavior-preserving; removes the larger of the two
copies and the drift risk between them.
