+++ packages/storage/src/sql/index.ts
@@ -7,5 +7,6 @@
 // organize-imports-ignore
 
 export * from "./Dialect";
+export * from "./mapPostgresType";
 export * from "./PredicateBuilder";
 export * from "./PrefixDdl";
